Font-size change to adjust to new monitor
Bigger monitor makes it easier to work with HiDPI and scaled fonts.
Overall XFCE can handle this fine, fonts in emacs and allacritty were
just too small.

This commit increases the font sizes to compensate for the 1.5x
scaling I am using on the workstation.

Still need to test how this will look in the laptop, though.